Rename BIO variables directly to variables
In the header data BIO variables are named but some variable names were also changed to what they actually represent. I now renamed the following BIO variables directly to what they represent:

- BIO1: Temprt
- BIO12: Precipt
- BIO7: Tem_range
- BIO15: Prec_Varieb

The reason is that everytime the data is read, those variables are renamed. This takes a lot of space and is unnecessary if they can just be renamed in the original data