Source preparation script
Instead of preparing the data in each script with the same code,
I used the preparation script instead and source it at the beginning.
I doublechecked all the resulting tables from the scripts and
they stay exactly the same. Therefore the new way of reading the
data has no effect on the results